Included Toeplitz Predictor Matrx
Changed the way the regression was calculated. Before, we applied a standard trace deconvolution formula each time we recorded an input. Now, we have a different column for each second after the impulse we wish to measure, and individually fit a coefficient to each of these columns ensuring that the trace deconvolution is precisely fitted by the algorithm and is not just approximated by a predetermined formula. This method is much more computationally intensive and will take a lot more time than the other method. However, the accuracy of the results can be seen to improve dramatically from a correlation ~ 0.35 to ~0.57.