Fully implemented reduced rank regression
Fully implemented reduced rank regression, still some questions on selecting an r-value for the rank to artificially reduce it and therefore probably reduce overfitting and noise. This script creates the Toeplitz predictor matrices for each stimulus, in the actual paper there are more stimuli used for consideration and each one has a specific time interval for which we create the Toeplitz whereas here we use the same interval for each stimulus and are currently only considering CS+, CS-, and licks.