added full matlab file support
included the ability to fully withhold a single stimulus from the regression model which we will be able to use in the future to assess the effects of each variable on the model. In this update, I mostly fix the issues with matlab file support as well as issues with the specific sample data used as it is recorded at 15 hertz. One can identify the sampling rate by looking at the Fall matlab file and looking at Fall['ops']['fs'].