Fully functioning withheld stimulus code
Sorted out the bugs with the withheld_stimulus functionality of the script. Now should be able to run either matlab (dictionary) data or 2d array data through the script to calculate the regression. There are still major issues with the performance of the model when applied to the new, dictionary data which I am still working on.